Add auth handlers

+++ src/main/java/com/example/carrental/dto/RentalResponseDto.java
@@ -3,6 +3,7 @@ package com.example.carrental.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -10,14 +11,9 @@ import java.time.LocalDate;
 public class RentalResponseDto {
 
     private Long id;
-
     private Long userId;
-
     private Long carId;
-
     private LocalDate startDate;
-
     private LocalDate endDate;
-
-    private Double totalPrice;
+    private BigDecimal totalPrice;
 }
